Extract core classes into library
