+++ BoothDownloadApp.Core/DatabaseManager.cs
