+++ BoothDownloadApp.Core/DownloadItem.cs
