+++ BoothDownloadApp.Core/DownloadManager.cs
