+++ BoothDownloadApp.Core/FilterManager.cs
