+++ BoothDownloadApp.Core/Settings.cs
