+++ BoothDownloadApp.sln
@@ -5,17 +5,23 @@ VisualStudioVersion = 17.13.35913.81 d17.13
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoothDownloadApp", "BoothDownloadApp\BoothDownloadApp.csproj", "{F806D4F7-E031-4240-8640-A72D2D59875F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoothDownloadApp.Core", "BoothDownloadApp.Core\BoothDownloadApp.Core.csproj", "{47A83B1D-BF3B-4A9F-BEE8-607B604B885D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F806D4F7-E031-4240-8640-A72D2D59875F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F806D4F7-E031-4240-8640-A72D2D59875F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F806D4F7-E031-4240-8640-A72D2D59875F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F806D4F7-E031-4240-8640-A72D2D59875F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
+        GlobalSection(ProjectConfigurationPlatforms) = postSolution
+                {F806D4F7-E031-4240-8640-A72D2D59875F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {F806D4F7-E031-4240-8640-A72D2D59875F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {F806D4F7-E031-4240-8640-A72D2D59875F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {F806D4F7-E031-4240-8640-A72D2D59875F}.Release|Any CPU.Build.0 = Release|Any CPU
+                {47A83B1D-BF3B-4A9F-BEE8-607B604B885D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {47A83B1D-BF3B-4A9F-BEE8-607B604B885D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {47A83B1D-BF3B-4A9F-BEE8-607B604B885D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {47A83B1D-BF3B-4A9F-BEE8-607B604B885D}.Release|Any CPU.Build.0 = Release|Any CPU
+        EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
