+++ BoothDownloadApp/MainWindow.xaml.cs
@@ -416,139 +416,4 @@ namespace BoothDownloadApp
         }
     }
 
-    public class BoothLibrary
-    {
-        [JsonPropertyName("library")]
-        public List<BoothItem> Library { get; set; } = new List<BoothItem>();
-    }
-
-
-    public class BoothItem : INotifyPropertyChanged
-    {
-        private bool _isSelected;
-        private string _description = string.Empty; // Initialize _description to a non-null value
-        private bool _isDownloaded;
-
-        [JsonPropertyName("productName")]
-        public string ProductName { get; set; } = string.Empty;
-
-        [JsonPropertyName("shopName")]
-        public string ShopName { get; set; } = string.Empty;
-
-        [JsonPropertyName("thumbnail")]
-        public string Thumbnail { get; set; } = string.Empty;
-
-        [JsonPropertyName("downloads")]
-        public List<DownloadInfo> Downloads { get; set; } = new List<DownloadInfo>();
-
-        // 商品単位の選択状態
-        public bool IsSelected
-        {
-            get => _isSelected;
-            set
-            {
-                if (_isSelected != value)
-                {
-                    _isSelected = value;
-                    OnPropertyChanged();
-                    if (Downloads != null)
-                    {
-                        foreach (var download in Downloads)
-                        {
-                            download.IsSelected = value;
-                        }
-                    }
-                }
-            }
-        }
-
-        [JsonIgnore]
-        public bool IsDownloaded
-        {
-            get => _isDownloaded;
-            set
-            {
-                if (_isDownloaded != value)
-                {
-                    _isDownloaded = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-
-        public event PropertyChangedEventHandler? PropertyChanged;
-
-        protected void OnPropertyChanged([CallerMemberName] string propertyName = null!)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        }
-
-        // 内部クラス DownloadInfo
-        public class DownloadInfo : INotifyPropertyChanged
-        {
-            private bool _isSelected;
-            private string _description = string.Empty; // Initialize _description to a non-null value
-            private bool _isDownloaded;
-
-            [JsonPropertyName("fileName")]
-            public string FileName { get; set; } = string.Empty;
-
-            [JsonPropertyName("downloadLink")]
-            public string DownloadLink { get; set; } = string.Empty;
-
-            [JsonPropertyName("description")]
-            public string Description
-            {
-                get => _description;
-                set
-                {
-                    if (_description != value)
-                    {
-                        _description = value;
-                        OnPropertyChanged();
-                    }
-                }
-            }
-
-            public bool IsSelected
-            {
-                get => _isSelected;
-                set
-                {
-                    if (_isSelected != value)
-                    {
-                        _isSelected = value;
-                        OnPropertyChanged();
-                    }
-                }
-            }
-
-            [JsonIgnore]
-            public bool IsDownloaded
-            {
-                get => _isDownloaded;
-                set
-                {
-                    if (_isDownloaded != value)
-                    {
-                        _isDownloaded = value;
-                        OnPropertyChanged();
-                    }
-                }
-            }
-
-            public event PropertyChangedEventHandler? PropertyChanged;
-
-            protected void OnPropertyChanged([CallerMemberName] string propertyName = null!)
-            {
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-            }
-        }
-
-        
-
-        
-    }
-
-
 }
